bugfix DELETE, UPDATE on User, Staff, Task

+++ src/main/java/com/netcracker/hotelbe/service/TaskService.java
@@ -30,6 +30,9 @@ public class TaskService {
     }
 
     public void deleteById(Long id){
+        if (!taskRepository.findById(id).isPresent()){
+            throw new EntityNotFoundException("No entity with id=" + id + " found");
+        }
         taskRepository.deleteById(id);
     }
